Fiscal Year and Finance Manage
Corrected the ability for the 'categories' input to post back to the
finance view the data so the success or failure tooltip would display
correctly. Set 'fiscalyear' to display in descending order. Set the
ordering of itmes to descending where it made logical sense - like
alphabetical order of categories and vendors in the form model/view.